refactor: add UserIcon
