+++ packages/widgets/src/HeaderBar/UserIcon/TextIcon.test.js
