+++ packages/widgets/src/HeaderBar/features/the_headerbar_contains_a_profile_menu/the_headerbar_shows_a_text_icon_if_the_user_does_not_have_an_avatar.js
@@ -1,6 +1,6 @@
 import '../common/index'
 import { Given, Then, When } from 'cypress-cucumber-preprocessor/steps'
-import { getInitials } from '../../TextIcon'
+import { getInitials } from '../../UserIcon/TextIcon.js'
 
 Given(
     'the HeaderBar loads without an error and the user does not have an avatar',
